InOut updates (#61)
This PR adds several InOut-related features+related tests in a rough fashion, to be cleaned up in a future PR.

* Allow connections between InOut ports on modules
* Allow Input/Output ports to connect to InOut ports
* Check for unconnected InOut ports (they can now be marked unused())

Note: direct connection (as opposed to assign-based connection) is being developed for InOut-related connections and will eventually replace assign-based connections in most places. That update is still a little while off, however - will involve updating most tests and consolidating a bunch of logic.