Allow ModDef::parameterize to take values larger than 32 bits (#93)
* Allow ModDef::parameterize to take values larger than 32 bits

This pulls in changes from slang-rs that lets us extract parameter
definitions from the SV file. Add a 'parameters' member to the module
definition that keeps the parameter definitions. Then, when
parameterizing, check these definitions to determine the width of the
literals that should be passed to slang and xlsynth.

* Allow ModDef::parameterize to take values larger than 32 bits

This pulls in changes from slang-rs that lets us extract parameter
definitions from the SV file. Add a 'parameters' member to the module
definition that keeps the parameter definitions. Then, when
parameterizing, check these definitions to determine the width of the
literals that should be passed to slang and xlsynth.

commit-id:03f2d387