Exclude vim swap files from spdx test (#96)
* Fix handling of parameters with dependent widths

The previous way of handling parameters would have an issue if the width
of one parameter depended on another, since the parameter types were
only parsed during loading of the original unparameterized module.
Remove the parameters field from ModDefCore and instead derive
the parameter types on demand when the module is re-parsed
with the new parameter values.

commit-id:0c00ed79

* Exclude vim swap files from spdx test

commit-id:b8462192